Split SqlServerFactCollector into per-domain partials
Behavior-neutral partial-class split: SqlServerFactCollector is now
`partial`, and its 28 Collect*FactsAsync methods are fanned out verbatim
into 6 domain files (Waits, Config, Resources, Storage, QueryPerf,
Activity). The ctor, fields, the CollectFactsAsync orchestrator, the
WS5 server-health constants, and BuildServerPropertiesQuery stay in the
base file. No call-site changes; method bodies are byte-identical to the
original.

Verified:
- dotnet build PerformanceMonitor.sln -c Debug: Build succeeded, 0 warnings, 0 errors
- Dashboard.Tests: 491 passed, 0 failed
- Lite.Tests: 545 passed, 0 failed
