+++ Dashboard/Analysis/SqlServerFactCollector.cs
@@ -14,7 +14,7 @@ namespace PerformanceMonitorDashboard.Analysis;
 /// Each fact category has its own collection method, added incrementally.
 /// Port of DuckDbFactCollector from Lite — queries collect.* tables instead of DuckDB views.
 /// </summary>
-public class SqlServerFactCollector : IFactCollector
+public partial class SqlServerFactCollector : IFactCollector
 {
     private readonly string _connectionString;
 
@@ -62,1909 +62,6 @@ public class SqlServerFactCollector : IFactCollector
     }
 
     /// <summary>
-    /// Collects wait stats facts — one Fact per significant wait type.
-    /// Value is wait_time_ms / period_duration_ms (fraction of examined period).
-    /// </summary>
-    private async Task CollectWaitStatsFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var command = connection.CreateCommand();
-            command.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    wait_type,
-    SUM(waiting_tasks_count_delta) AS total_waiting_tasks,
-    SUM(wait_time_ms_delta) AS total_wait_time_ms,
-    SUM(signal_wait_time_ms_delta) AS total_signal_wait_time_ms
-FROM collect.wait_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime
-AND   wait_time_ms_delta > 0
-GROUP BY wait_type
-ORDER BY SUM(wait_time_ms_delta) DESC";
-
-            command.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            command.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await command.ExecuteReaderAsync();
-            while (await reader.ReadAsync())
-            {
-                var waitType = reader.GetString(0);
-                var waitingTasks = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-                var waitTimeMs = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-                var signalWaitTimeMs = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-
-                if (waitTimeMs <= 0) continue;
-
-                var fractionOfPeriod = waitTimeMs / context.PeriodDurationMs;
-                var avgMsPerWait = waitingTasks > 0 ? (double)waitTimeMs / waitingTasks : 0;
-
-                facts.Add(new Fact
-                {
-                    Source = "waits",
-                    Key = waitType,
-                    Value = fractionOfPeriod,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["wait_time_ms"] = waitTimeMs,
-                        ["waiting_tasks_count"] = waitingTasks,
-                        ["signal_wait_time_ms"] = signalWaitTimeMs,
-                        ["resource_wait_time_ms"] = waitTimeMs - signalWaitTimeMs,
-                        ["avg_ms_per_wait"] = avgMsPerWait,
-                        ["period_duration_ms"] = context.PeriodDurationMs
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectWaitStatsFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects blocking facts from blocking_BlockedProcessReport.
-    /// Produces a single BLOCKING_EVENTS fact with event count, rate, and details.
-    /// Value is events per hour for threshold comparison.
-    /// </summary>
-    private async Task CollectBlockingFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var command = connection.CreateCommand();
-            command.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    COUNT(*) AS event_count,
-    AVG(CAST(wait_time_ms AS FLOAT)) AS avg_wait_time_ms,
-    MAX(wait_time_ms) AS max_wait_time_ms,
-    COUNT(DISTINCT spid) AS distinct_head_blockers,
-    COUNT(CASE WHEN status = 'sleeping' THEN 1 END) AS sleeping_blocker_count
-FROM collect.blocking_BlockedProcessReport
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            command.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            command.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await command.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var eventCount = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (eventCount <= 0) return;
-
-            var avgWaitTimeMs = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var maxWaitTimeMs = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var distinctHeadBlockers = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var sleepingBlockerCount = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            var periodHours = context.PeriodDurationMs / 3_600_000.0;
-            var eventsPerHour = periodHours > 0 ? eventCount / periodHours : 0;
-
-            facts.Add(new Fact
-            {
-                Source = "blocking",
-                Key = "BLOCKING_EVENTS",
-                Value = eventsPerHour,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["event_count"] = eventCount,
-                    ["events_per_hour"] = eventsPerHour,
-                    ["avg_wait_time_ms"] = avgWaitTimeMs,
-                    ["max_wait_time_ms"] = maxWaitTimeMs,
-                    ["distinct_head_blockers"] = distinctHeadBlockers,
-                    ["sleeping_blocker_count"] = sleepingBlockerCount,
-                    ["period_hours"] = periodHours
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectBlockingFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects deadlock facts from the deadlocks table.
-    /// Produces a single DEADLOCKS fact with count and rate.
-    /// Value is deadlocks per hour for threshold comparison.
-    /// </summary>
-    private async Task CollectDeadlockFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var command = connection.CreateCommand();
-            command.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT COUNT(*) AS deadlock_count
-FROM collect.deadlocks
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            command.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            command.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await command.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var deadlockCount = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (deadlockCount <= 0) return;
-
-            var periodHours = context.PeriodDurationMs / 3_600_000.0;
-            var deadlocksPerHour = periodHours > 0 ? deadlockCount / periodHours : 0;
-
-            facts.Add(new Fact
-            {
-                Source = "blocking",
-                Key = "DEADLOCKS",
-                Value = deadlocksPerHour,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["deadlock_count"] = deadlockCount,
-                    ["deadlocks_per_hour"] = deadlocksPerHour,
-                    ["period_hours"] = periodHours
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectDeadlockFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects server configuration settings relevant to analysis.
-    /// These become facts that amplifiers and the config audit tool can reference
-    /// to make recommendations specific (e.g., "your CTFP is 50" vs "check CTFP").
-    /// </summary>
-    private async Task CollectServerConfigFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            // Latest value PER configuration_name (ROW_NUMBER, not TOP N): server_configuration_history
-            // accumulates a row per collection, so a naive TOP-N-ORDER-BY-time returns the newest N
-            // ROWS — which collapses to one config when collections are frequent, silently dropping
-            // settings. Partition by name and take rn = 1 so each requested setting is its latest value.
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        configuration_name,
-        CAST(value_in_use AS BIGINT) AS value_in_use,
-        ROW_NUMBER() OVER (PARTITION BY configuration_name ORDER BY collection_time DESC) AS rn
-    FROM config.server_configuration_history
-    WHERE configuration_name IN (
-        'cost threshold for parallelism',
-        'max degree of parallelism',
-        'max server memory (MB)',
-        'min server memory (MB)',
-        'max worker threads'
-    )
-)
-SELECT
-    configuration_name,
-    value_in_use
-FROM latest
-WHERE rn = 1";
-
-            // max/min server memory are read alongside the rooted CONFIG_* facts so the
-            // narrow-memory derivation below can compare them without a second query.
-            double? maxMemoryMb = null;
-            double? minMemoryMb = null;
-
-            using (var reader = await cmd.ExecuteReaderAsync())
-            {
-                while (await reader.ReadAsync())
-                {
-                    var configName = reader.GetString(0);
-                    var value = Convert.ToDouble(reader.GetValue(1));
-
-                    if (configName == "max server memory (MB)") maxMemoryMb = value;
-                    if (configName == "min server memory (MB)") minMemoryMb = value;
-
-                    var factKey = configName switch
-                    {
-                        "cost threshold for parallelism" => "CONFIG_CTFP",
-                        "max degree of parallelism" => "CONFIG_MAXDOP",
-                        "max server memory (MB)" => "CONFIG_MAX_MEMORY_MB",
-                        "min server memory (MB)" => "CONFIG_MIN_MEMORY_MB",
-                        "max worker threads" => "CONFIG_MAX_WORKER_THREADS",
-                        _ => null
-                    };
-
-                    if (factKey == null) continue;
-
-                    facts.Add(new Fact
-                    {
-                        Source = "config",
-                        Key = factKey,
-                        Value = value,
-                        ServerId = context.ServerId,
-                        Metadata = new Dictionary<string, double>
-                        {
-                            ["value_in_use"] = value
-                        }
-                    });
-                }
-            }
-
-            // CONFIG_MIN_MAX_MEMORY_NARROW: emitted only when max is configured AND min is pinned
-            // near it (shared rule so Dashboard/Lite agree).
-            var narrow = FactRemediation.BuildNarrowMemoryFact(context.ServerId, maxMemoryMb, minMemoryMb);
-            if (narrow is not null)
-                facts.Add(narrow);
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectServerConfigFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects memory stats: total physical RAM, buffer pool size, target memory.
-    /// These facts enable edition-aware memory recommendations in the config audit.
-    /// </summary>
-    private async Task CollectMemoryFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT TOP 1
-    total_physical_memory_mb,
-    buffer_pool_mb,
-    committed_target_memory_mb
-FROM collect.memory_stats
-WHERE collection_time <= @endTime
-ORDER BY collection_time DESC";
-
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalPhysical = reader.IsDBNull(0) ? 0.0 : Convert.ToDouble(reader.GetValue(0));
-            var bufferPool = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var targetMemory = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-
-            if (totalPhysical > 0)
-                facts.Add(new Fact { Source = "memory", Key = "MEMORY_TOTAL_PHYSICAL_MB", Value = totalPhysical, ServerId = context.ServerId });
-            if (bufferPool > 0)
-                facts.Add(new Fact { Source = "memory", Key = "MEMORY_BUFFER_POOL_MB", Value = bufferPool, ServerId = context.ServerId });
-            if (targetMemory > 0)
-                facts.Add(new Fact { Source = "memory", Key = "MEMORY_TARGET_MB", Value = targetMemory, ServerId = context.ServerId });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectMemoryFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects total database data size from file_io_stats.
-    /// Sums the latest size_on_disk_bytes across all database files for the server.
-    /// </summary>
-    private async Task CollectDatabaseSizeFactAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        database_name,
-        file_name,
-        size_on_disk_bytes,
-        ROW_NUMBER() OVER (PARTITION BY database_name, file_name ORDER BY collection_time DESC) AS rn
-    FROM collect.file_io_stats
-    WHERE collection_time <= @endTime
-    AND   size_on_disk_bytes > 0
-)
-SELECT SUM(size_on_disk_bytes / 1048576.0) AS total_size_mb
-FROM latest
-WHERE rn = 1";
-
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalSize = reader.IsDBNull(0) ? 0.0 : Convert.ToDouble(reader.GetValue(0));
-            if (totalSize > 0)
-                facts.Add(new Fact { Source = "config", Key = "DATABASE_TOTAL_SIZE_MB", Value = totalSize, ServerId = context.ServerId });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectDatabaseSizeFactAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects SQL Server edition and major version from the server_properties table.
-    /// </summary>
-    private async Task CollectServerMetadataFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT TOP 1
-    engine_edition,
-    CAST(LEFT(product_version, CHARINDEX('.', product_version) - 1) AS INT) AS major_version
-FROM collect.server_properties
-ORDER BY collection_time DESC";
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var edition = reader.IsDBNull(0) ? 0 : Convert.ToInt32(reader.GetValue(0));
-            var majorVersion = reader.IsDBNull(1) ? 0 : Convert.ToInt32(reader.GetValue(1));
-
-            if (edition > 0)
-                facts.Add(new Fact { Source = "config", Key = "SERVER_EDITION", Value = edition, ServerId = context.ServerId });
-            if (majorVersion > 0)
-                facts.Add(new Fact { Source = "config", Key = "SERVER_MAJOR_VERSION", Value = majorVersion, ServerId = context.ServerId });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectServerMetadataFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects CPU utilization: average and max SQL Server CPU % over the period.
-    /// Value is average SQL CPU %. Corroborates SOS_SCHEDULER_YIELD.
-    /// </summary>
-    private async Task CollectCpuUtilizationFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    AVG(CAST(sqlserver_cpu_utilization AS FLOAT)) AS avg_sql_cpu,
-    MAX(sqlserver_cpu_utilization) AS max_sql_cpu,
-    AVG(CAST(other_process_cpu_utilization AS FLOAT)) AS avg_other_cpu,
-    MAX(other_process_cpu_utilization) AS max_other_cpu,
-    COUNT(*) AS sample_count
-FROM collect.cpu_utilization_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var avgSqlCpu = reader.IsDBNull(0) ? 0.0 : Convert.ToDouble(reader.GetValue(0));
-            var maxSqlCpu = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var avgOtherCpu = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-            var maxOtherCpu = reader.IsDBNull(3) ? 0.0 : Convert.ToDouble(reader.GetValue(3));
-            var sampleCount = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            if (sampleCount == 0) return;
-
-            var cpuMetadata = new Dictionary<string, double>
-            {
-                ["avg_sql_cpu"] = avgSqlCpu,
-                ["max_sql_cpu"] = maxSqlCpu,
-                ["avg_other_cpu"] = avgOtherCpu,
-                ["max_other_cpu"] = maxOtherCpu,
-                ["avg_total_cpu"] = avgSqlCpu + avgOtherCpu,
-                ["sample_count"] = sampleCount
-            };
-
-            facts.Add(new Fact
-            {
-                Source = "cpu",
-                Key = "CPU_SQL_PERCENT",
-                Value = avgSqlCpu,
-                ServerId = context.ServerId,
-                Metadata = cpuMetadata
-            });
-
-            // Emit a CPU_SPIKE fact when max is high and significantly above average.
-            // This catches bursty CPU events that average-based scoring misses entirely.
-            // Requires max >= 80% AND at least 3x the average (or avg < 20% with max >= 80%).
-            if (maxSqlCpu >= 80 && (avgSqlCpu < 20 || maxSqlCpu / Math.Max(avgSqlCpu, 1) >= 3))
-            {
-                facts.Add(new Fact
-                {
-                    Source = "cpu",
-                    Key = "CPU_SPIKE",
-                    Value = maxSqlCpu,
-                    ServerId = context.ServerId,
-                    Metadata = cpuMetadata
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectCpuUtilizationFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects I/O latency from file_io_stats delta columns.
-    /// Computes average read and write latency across all database files.
-    /// </summary>
-    private async Task CollectIoLatencyFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    SUM(io_stall_read_ms_delta) AS total_stall_read_ms,
-    SUM(num_of_reads_delta) AS total_reads,
-    SUM(io_stall_write_ms_delta) AS total_stall_write_ms,
-    SUM(num_of_writes_delta) AS total_writes
-FROM collect.file_io_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime
-AND   (num_of_reads_delta > 0 OR num_of_writes_delta > 0)";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalStallReadMs = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            var totalReads = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var totalStallWriteMs = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalWrites = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-
-            if (totalReads > 0)
-            {
-                var avgReadLatency = (double)totalStallReadMs / totalReads;
-                facts.Add(new Fact
-                {
-                    Source = "io",
-                    Key = "IO_READ_LATENCY_MS",
-                    Value = avgReadLatency,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["avg_read_latency_ms"] = avgReadLatency,
-                        ["total_stall_read_ms"] = totalStallReadMs,
-                        ["total_reads"] = totalReads
-                    }
-                });
-            }
-
-            if (totalWrites > 0)
-            {
-                var avgWriteLatency = (double)totalStallWriteMs / totalWrites;
-                facts.Add(new Fact
-                {
-                    Source = "io",
-                    Key = "IO_WRITE_LATENCY_MS",
-                    Value = avgWriteLatency,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["avg_write_latency_ms"] = avgWriteLatency,
-                        ["total_stall_write_ms"] = totalStallWriteMs,
-                        ["total_writes"] = totalWrites
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectIoLatencyFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects TempDB usage facts: max usage, version store size, and unallocated space.
-    /// Value is max total_reserved_mb over the period.
-    /// Dashboard uses computed columns (total_reserved_mb, etc.) from collect.tempdb_stats.
-    /// </summary>
-    private async Task CollectTempDbFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    MAX(total_reserved_mb) AS max_total_reserved_mb,
-    MAX(user_object_reserved_mb) AS max_user_object_mb,
-    MAX(internal_object_reserved_mb) AS max_internal_object_mb,
-    MAX(version_store_reserved_mb) AS max_version_store_mb,
-    MIN(unallocated_mb) AS min_unallocated_mb,
-    AVG(CAST(total_reserved_mb AS FLOAT)) AS avg_total_reserved_mb
-FROM collect.tempdb_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var maxReserved = reader.IsDBNull(0) ? 0.0 : Convert.ToDouble(reader.GetValue(0));
-            var maxUserObj = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var maxInternalObj = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-            var maxVersionStore = reader.IsDBNull(3) ? 0.0 : Convert.ToDouble(reader.GetValue(3));
-            var minUnallocated = reader.IsDBNull(4) ? 0.0 : Convert.ToDouble(reader.GetValue(4));
-            var avgReserved = reader.IsDBNull(5) ? 0.0 : Convert.ToDouble(reader.GetValue(5));
-
-            if (maxReserved <= 0) return;
-
-            // TempDB usage as fraction of total space (reserved + unallocated)
-            var totalSpace = maxReserved + minUnallocated;
-            var usageFraction = totalSpace > 0 ? maxReserved / totalSpace : 0;
-
-            facts.Add(new Fact
-            {
-                Source = "tempdb",
-                Key = "TEMPDB_USAGE",
-                Value = usageFraction,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["max_reserved_mb"] = maxReserved,
-                    ["avg_reserved_mb"] = avgReserved,
-                    ["max_user_object_mb"] = maxUserObj,
-                    ["max_internal_object_mb"] = maxInternalObj,
-                    ["max_version_store_mb"] = maxVersionStore,
-                    ["min_unallocated_mb"] = minUnallocated,
-                    ["usage_fraction"] = usageFraction
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectTempDbFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects memory grant facts from the memory_grant_stats table.
-    /// Detects grant waiters (sessions waiting for memory) and grant pressure.
-    /// </summary>
-    private async Task CollectMemoryGrantFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    MAX(waiter_count) AS max_waiters,
-    AVG(CAST(waiter_count AS FLOAT)) AS avg_waiters,
-    MAX(grantee_count) AS max_grantees,
-    SUM(timeout_error_count_delta) AS total_timeout_errors,
-    SUM(forced_grant_count_delta) AS total_forced_grants
-FROM collect.memory_grant_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var maxWaiters = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            var avgWaiters = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var maxGrantees = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalTimeouts = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var totalForcedGrants = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            // Only create a fact if there's evidence of grant pressure
-            if (maxWaiters <= 0 && totalTimeouts <= 0 && totalForcedGrants <= 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "memory",
-                Key = "MEMORY_GRANT_PENDING",
-                Value = maxWaiters,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["max_waiters"] = maxWaiters,
-                    ["avg_waiters"] = avgWaiters,
-                    ["max_grantees"] = maxGrantees,
-                    ["total_timeout_errors"] = totalTimeouts,
-                    ["total_forced_grants"] = totalForcedGrants
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectMemoryGrantFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects query-level aggregate facts from query_stats.
-    /// Focuses on spills (memory grant misestimates) and high-parallelism queries.
-    /// </summary>
-    private async Task CollectQueryStatsFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    SUM(total_spills) AS total_spills,
-    COUNT(CASE WHEN max_dop > 8 THEN 1 END) AS high_dop_queries,
-    COUNT(CASE WHEN total_spills > 0 THEN 1 END) AS spilling_queries,
-    SUM(execution_count_delta) AS total_executions,
-    SUM(total_worker_time_delta) AS total_cpu_time_us
-FROM collect.query_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime
-AND   execution_count_delta > 0";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalSpills = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            var highDopQueries = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var spillingQueries = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalExecutions = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var totalCpuTimeUs = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            if (totalSpills > 0)
-            {
-                facts.Add(new Fact
-                {
-                    Source = "queries",
-                    Key = "QUERY_SPILLS",
-                    Value = totalSpills,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["total_spills"] = totalSpills,
-                        ["spilling_query_count"] = spillingQueries,
-                        ["total_executions"] = totalExecutions
-                    }
-                });
-            }
-
-            if (highDopQueries > 0)
-            {
-                facts.Add(new Fact
-                {
-                    Source = "queries",
-                    Key = "QUERY_HIGH_DOP",
-                    Value = highDopQueries,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["high_dop_query_count"] = highDopQueries,
-                        ["total_cpu_time_us"] = totalCpuTimeUs,
-                        ["total_executions"] = totalExecutions
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectQueryStatsFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Detects parameter-sensitive cached plans: a single query_plan_hash whose
-    /// per-execution worker time varies wildly — one plan serving very different
-    /// parameter values. Emits one aggregate PARAMETER_SENSITIVITY fact.
-    /// Note min_*/max_* are cumulative over the plan's cached lifetime, so the
-    /// finding means "this plan, active now, has a history of widely varying cost".
-    /// </summary>
-    private async Task CollectParameterSensitivityFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-WITH latest AS
-(
-    SELECT
-        query_hash,
-        query_plan_hash,
-        database_name,
-        execution_count,
-        creation_time,
-        min_worker_time,
-        max_worker_time,
-        min_grant_kb,
-        max_grant_kb,
-        min_spills,
-        max_spills,
-        ROW_NUMBER() OVER
-        (
-            PARTITION BY database_name, query_hash, query_plan_hash
-            ORDER BY collection_time DESC
-        ) AS rn
-    FROM collect.query_stats
-    WHERE collection_time >= @startTime
-    AND   collection_time <= @endTime
-    AND   execution_count_delta > 0
-)
-SELECT
-    min_worker_time,
-    max_worker_time,
-    CAST(max_worker_time AS float) / NULLIF(min_worker_time, 0) AS worker_ratio,
-    CAST(max_grant_kb AS float) / NULLIF(min_grant_kb, 0) AS grant_ratio,
-    CASE WHEN max_spills > 0 AND min_spills = 0 THEN 1 ELSE 0 END AS spill_divergence
-FROM latest
-WHERE rn = 1
-AND   min_worker_time >= 10000
-AND   max_worker_time >= 250000
-AND   execution_count >= 20
-AND   creation_time <= @startTime
-AND   CAST(max_worker_time AS float) / NULLIF(min_worker_time, 0) >= 10
-ORDER BY worker_ratio DESC
-OFFSET 0 ROWS FETCH NEXT 20 ROWS ONLY";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            var offenderCount = 0;
-            var worstRatio = 0.0;
-            var worstMinWorker = 0L;
-            var worstMaxWorker = 0L;
-            var worstGrantRatio = 0.0;
-            var worstSpillDivergence = 0;
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            while (await reader.ReadAsync())
-            {
-                // Rows arrive ordered by worker_ratio DESC — the first row is the worst offender.
-                if (offenderCount == 0)
-                {
-                    worstMinWorker = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-                    worstMaxWorker = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-                    worstRatio = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-                    worstGrantRatio = reader.IsDBNull(3) ? 0.0 : Convert.ToDouble(reader.GetValue(3));
-                    worstSpillDivergence = reader.IsDBNull(4) ? 0 : Convert.ToInt32(reader.GetValue(4));
-                }
-                offenderCount++;
-            }
-
-            if (offenderCount == 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "queries",
-                Key = "PARAMETER_SENSITIVITY",
-                Value = worstRatio,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["offender_count"] = offenderCount,
-                    ["worst_ratio"] = worstRatio,
-                    ["worst_min_worker_us"] = worstMinWorker,
-                    ["worst_max_worker_us"] = worstMaxWorker,
-                    ["worst_grant_ratio"] = worstGrantRatio,
-                    ["grant_divergence"] = worstGrantRatio >= 5 ? 1 : 0,
-                    ["spill_divergence"] = worstSpillDivergence
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectParameterSensitivityFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Detects plan regressions: a query whose currently-active plan has per-execution
-    /// cost >= 2x the best plan that query is known to perform well with. Emits one
-    /// aggregate PLAN_REGRESSION fact. Sourced from Query Store (collect.query_store_data);
-    /// no fact when Query Store is not enabled on the monitored databases.
-    /// Unlike other collectors this windows on server_last_execution_time (14-day
-    /// comparison window), NOT collection_time.
-    /// </summary>
-    private async Task CollectPlanRegressionFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-WITH deduped AS
-(
-    -- Collapse incremental re-collections of the same open runtime-stats interval:
-    -- keep only the latest collection_time row per logical interval.
-    SELECT
-        database_name,
-        query_id,
-        plan_id,
-        query_plan_hash,
-        count_executions,
-        avg_cpu_time,
-        avg_duration,
-        server_last_execution_time,
-        is_forced_plan,
-        force_failure_count,
-        ROW_NUMBER() OVER
-        (
-            PARTITION BY database_name, query_id, plan_id, server_first_execution_time
-            ORDER BY collection_time DESC
-        ) AS rn
-    FROM collect.query_store_data
-    WHERE execution_type_desc = N'Regular'
-    AND   server_last_execution_time >= @windowStart
-),
-plan_agg AS
-(
-    -- Execution-weighted per-exec cost per plan_id. query_plan_hash is invariant
-    -- within a plan_id, so include it in the GROUP BY rather than aggregating it
-    -- (MS Learn's MAX page does not list binary/varbinary in the accepted types).
-    SELECT
-        database_name,
-        query_id,
-        plan_id,
-        query_plan_hash,
-        SUM(count_executions) AS execs,
-        CASE WHEN SUM(count_executions) > 0
-             THEN SUM(avg_cpu_time * count_executions) / NULLIF(SUM(count_executions), 0)
-             ELSE 0 END AS cpu_per_exec,
-        CASE WHEN SUM(count_executions) > 0
-             THEN SUM(avg_duration * count_executions) / NULLIF(SUM(count_executions), 0)
-             ELSE 0 END AS dur_per_exec,
-        MAX(server_last_execution_time) AS last_exec,
-        MAX(CAST(is_forced_plan AS tinyint)) AS is_forced_plan,
-        MAX(force_failure_count) AS force_failure_count
-    FROM deduped
-    WHERE rn = 1
-    GROUP BY database_name, query_id, plan_id, query_plan_hash
-),
-plan_dedup AS
-(
-    -- Collapse plan_ids that share a query_plan_hash (a recompile can produce an
-    -- identical plan under a new plan_id); keep only plans with enough executions.
-    SELECT
-        database_name,
-        query_id,
-        query_plan_hash,
-        SUM(execs) AS execs,
-        CASE WHEN SUM(execs) > 0
-             THEN SUM(cpu_per_exec * execs) / NULLIF(SUM(execs), 0)
-             ELSE 0 END AS cpu_per_exec,
-        CASE WHEN SUM(execs) > 0
-             THEN SUM(dur_per_exec * execs) / NULLIF(SUM(execs), 0)
-             ELSE 0 END AS dur_per_exec,
-        MAX(last_exec) AS last_exec,
-        MAX(is_forced_plan) AS is_forced_plan,
-        MAX(force_failure_count) AS force_failure_count
-    FROM plan_agg
-    GROUP BY database_name, query_id, query_plan_hash
-    HAVING SUM(execs) >= 25
-),
-ranked AS
-(
-    SELECT
-        *,
-        ROW_NUMBER() OVER (PARTITION BY database_name, query_id ORDER BY last_exec DESC) AS recency,
-        ROW_NUMBER() OVER (PARTITION BY database_name, query_id ORDER BY cpu_per_exec ASC) AS cheapness
-    FROM plan_dedup
-),
-compared AS
-(
-    -- Latest active plan vs the best-performing plan for the same query.
-    SELECT
-        l.query_id,
-        l.cpu_per_exec AS latest_cpu,
-        l.dur_per_exec AS latest_dur,
-        l.is_forced_plan AS latest_is_forced,
-        l.force_failure_count AS force_failure_count,
-        b.cpu_per_exec AS best_cpu,
-        b.dur_per_exec AS best_dur,
-        (SELECT MAX(v)
-         FROM (VALUES
-             (CAST(l.cpu_per_exec AS float) / NULLIF(b.cpu_per_exec, 0)),
-             (CAST(l.dur_per_exec AS float) / NULLIF(b.dur_per_exec, 0))
-         ) AS x(v)) AS regression_factor
-    FROM ranked AS l
-    JOIN ranked AS b
-      ON  b.database_name = l.database_name
-      AND b.query_id = l.query_id
-      AND b.cheapness = 1
-    WHERE l.recency = 1
-    AND   l.query_plan_hash <> b.query_plan_hash
-)
-SELECT
-    query_id,
-    latest_cpu,
-    latest_dur,
-    latest_is_forced,
-    force_failure_count,
-    best_cpu,
-    best_dur,
-    regression_factor
-FROM compared
-WHERE regression_factor >= 2
-ORDER BY regression_factor DESC
-OFFSET 0 ROWS FETCH NEXT 20 ROWS ONLY";
-
-            cmd.Parameters.Add(new SqlParameter("@windowStart", context.TimeRangeStart.AddDays(-14)));
-
-            var offenderCount = 0;
-            var worstFactor = 0.0;
-            var worstQueryId = 0L;
-            var worstLatestCpu = 0.0;
-            var worstBestCpu = 0.0;
-            var worstDimension = 1;
-            var worstLatestForced = 0;
-            var worstForceFailures = 0L;
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            while (await reader.ReadAsync())
-            {
-                // Rows arrive ordered by regression_factor DESC — the first row is the worst offender.
-                if (offenderCount == 0)
-                {
-                    worstQueryId = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-                    var latestCpu = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-                    var latestDur = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-                    worstLatestForced = (!reader.IsDBNull(3) && Convert.ToInt32(reader.GetValue(3)) > 0) ? 1 : 0;
-                    worstForceFailures = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-                    var bestCpu = reader.IsDBNull(5) ? 0.0 : Convert.ToDouble(reader.GetValue(5));
-                    var bestDur = reader.IsDBNull(6) ? 0.0 : Convert.ToDouble(reader.GetValue(6));
-                    worstFactor = reader.IsDBNull(7) ? 0.0 : Convert.ToDouble(reader.GetValue(7));
-
-                    worstLatestCpu = latestCpu;
-                    worstBestCpu = bestCpu;
-                    var cpuRatio = bestCpu > 0 ? latestCpu / bestCpu : 0.0;
-                    var durRatio = bestDur > 0 ? latestDur / bestDur : 0.0;
-                    worstDimension = cpuRatio >= durRatio ? 1 : 2; // 1 = cpu, 2 = duration
-                }
-                offenderCount++;
-            }
-
-            if (offenderCount == 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "queries",
-                Key = "PLAN_REGRESSION",
-                Value = worstFactor,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["offender_count"] = offenderCount,
-                    ["worst_regression_factor"] = worstFactor,
-                    ["worst_query_id"] = worstQueryId,
-                    ["latest_cpu_per_exec_us"] = worstLatestCpu,
-                    ["best_cpu_per_exec_us"] = worstBestCpu,
-                    ["regressed_dimension"] = worstDimension,
-                    ["latest_is_forced"] = worstLatestForced,
-                    ["force_failure_count"] = worstForceFailures
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectPlanRegressionFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Reconstructs blocking chains from collect.blocking_BlockedProcessReport (one row
-    /// per side of each blocking event, dedup'd to <c>activity = 'blocked'</c>) and emits
-    /// one aggregate BLOCKING_CHAIN fact describing the worst chain — apex head blocker,
-    /// depth, transitive victim count. Structure the BLOCKING_EVENTS rate is blind to.
-    /// Reads typed blocker-side columns populated by collect.process_blocked_process_xml,
-    /// so no XML re-parse on the analysis hot path.
-    /// </summary>
-    private async Task CollectBlockingChainFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        const int maxPairs = 5000;
-        const int maxDepth = 50;
-        const int stepBudget = 100_000;
-
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            // ORDER BY collection_time DESC is a backward CIX scan (sort-free); event_time
-            // is a residual predicate. activity='blocked' picks the canonical per-event side.
-            // blocking_spid IS NOT NULL filters out rows whose source XML had an empty
-            // <blocking-process><process/></blocking-process> (system task / torn-down session) —
-            // those can't contribute to a reconstructed chain.
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT TOP (5000)
-    event_time,
-    database_name,
-    spid,
-    last_transaction_started,
-    blocking_spid,
-    blocking_last_tran_started,
-    wait_time_ms,
-    lock_mode,
-    blocking_status,
-    blocked_sql_text,
-    blocking_sql_text
-FROM collect.blocking_BlockedProcessReport
-WHERE collection_time >= @collectionWindow
-AND   event_time >= @startTime
-AND   event_time <= @endTime
-AND   activity = 'blocked'
-AND   blocking_spid IS NOT NULL
-ORDER BY collection_time DESC";
-
-            // Generous bound — analysis window plus an hour — to catch rows whose
-            // event_time is inside the window but whose collection_time may lag slightly.
-            cmd.Parameters.Add(new SqlParameter("@collectionWindow", context.TimeRangeStart.AddHours(-1)));
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            var rows = new List<BlockingPairRow>();
-            using (var reader = await cmd.ExecuteReaderAsync())
-            {
-                while (await reader.ReadAsync())
-                {
-                    rows.Add(new BlockingPairRow
-                    {
-                        EventTime = reader.IsDBNull(0) ? default : reader.GetDateTime(0),
-                        DatabaseName = reader.IsDBNull(1) ? string.Empty : reader.GetString(1),
-                        BlockedSpid = reader.IsDBNull(2) ? 0 : Convert.ToInt32(reader.GetValue(2)),
-                        BlockedTranStarted = reader.IsDBNull(3) ? (DateTime?)null : reader.GetDateTime(3),
-                        BlockingSpid = reader.IsDBNull(4) ? 0 : Convert.ToInt32(reader.GetValue(4)),
-                        BlockingTranStarted = reader.IsDBNull(5) ? (DateTime?)null : reader.GetDateTime(5),
-                        WaitTimeMs = reader.IsDBNull(6) ? 0L : Convert.ToInt64(reader.GetValue(6)),
-                        LockMode = reader.IsDBNull(7) ? string.Empty : reader.GetString(7),
-                        BlockingStatus = reader.IsDBNull(8) ? string.Empty : reader.GetString(8),
-                        BlockedSqlText = reader.IsDBNull(9) ? string.Empty : reader.GetString(9),
-                        BlockingSqlText = reader.IsDBNull(10) ? string.Empty : reader.GetString(10)
-                    });
-                }
-            }
-
-            if (rows.Count == 0) return;
-
-            var reconstruction = BlockingChainReconstructor.Reconstruct(rows, maxDepth, maxPairs, stepBudget);
-            if (reconstruction.Chains.Count == 0) return;
-
-            var worst = reconstruction.Chains[0];
-
-            facts.Add(new Fact
-            {
-                Source = "blocking",
-                Key = "BLOCKING_CHAIN",
-                Value = worst.Depth,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["worst_chain_depth"] = worst.Depth,
-                    ["worst_chain_victim_count"] = worst.VictimCount,
-                    ["worst_apex_spid"] = worst.ApexSpid,
-                    ["worst_apex_sleeping"] = worst.ApexSleeping ? 1 : 0,
-                    ["worst_chain_max_wait_ms"] = worst.MaxWaitMs,
-                    ["total_reconstructed_chains"] = reconstruction.Chains.Count,
-                    ["deepest_chain_overall"] = reconstruction.Chains.Max(c => c.Depth),
-                    ["max_victim_count_overall"] = reconstruction.Chains.Max(c => c.VictimCount),
-                    ["depth_capped"] = reconstruction.DepthCapped ? 1 : 0,
-                    ["traversal_truncated"] = reconstruction.TraversalTruncated ? 1 : 0,
-                    ["cycle_detected"] = reconstruction.CycleDetected ? 1 : 0
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectBlockingChainFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Identifies individual queries that are consistently terrible ("bad actors").
-    /// These queries don't necessarily cause server-level symptoms but waste resources
-    /// on every execution. Detection uses execution count tiers x per-execution impact.
-    /// Top 5 worst offenders become individual BAD_ACTOR facts.
-    /// Dashboard query_hash is binary(8) — convert to hex string for fact key.
-    /// </summary>
-    private async Task CollectBadActorFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT TOP 5
-    database_name,
-    CONVERT(VARCHAR(18), query_hash, 1) AS query_hash,
-    CAST(SUM(execution_count_delta) AS BIGINT) AS exec_count,
-    CASE WHEN SUM(execution_count_delta) > 0
-         THEN CAST(SUM(total_worker_time_delta) AS FLOAT) / SUM(execution_count_delta) / 1000.0
-         ELSE 0 END AS avg_cpu_ms,
-    CASE WHEN SUM(execution_count_delta) > 0
-         THEN CAST(SUM(total_elapsed_time_delta) AS FLOAT) / SUM(execution_count_delta) / 1000.0
-         ELSE 0 END AS avg_elapsed_ms,
-    CASE WHEN SUM(execution_count_delta) > 0
-         THEN CAST(SUM(total_logical_reads_delta) AS FLOAT) / SUM(execution_count_delta)
-         ELSE 0 END AS avg_reads,
-    CAST(SUM(total_worker_time_delta) AS BIGINT) AS total_cpu_us,
-    CAST(SUM(total_logical_reads_delta) AS BIGINT) AS total_reads,
-    CAST(SUM(total_spills) AS BIGINT) AS total_spills,
-    MAX(max_dop) AS max_dop,
-    LEFT(CAST(DECOMPRESS(MAX(query_text)) AS NVARCHAR(MAX)), 200) AS query_text
-FROM collect.query_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime
-AND   execution_count_delta > 0
-GROUP BY database_name, query_hash
-HAVING SUM(execution_count_delta) >= 100
-ORDER BY CAST(SUM(total_worker_time_delta) AS FLOAT) / NULLIF(SUM(execution_count_delta), 0) *
-         LOG(NULLIF(SUM(execution_count_delta), 0)) DESC";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            while (await reader.ReadAsync())
-            {
-                var dbName = reader.IsDBNull(0) ? "" : reader.GetString(0);
-                var queryHash = reader.IsDBNull(1) ? "" : reader.GetString(1);
-                var execCount = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-                var avgCpuMs = reader.IsDBNull(3) ? 0.0 : Convert.ToDouble(reader.GetValue(3));
-                var avgElapsedMs = reader.IsDBNull(4) ? 0.0 : Convert.ToDouble(reader.GetValue(4));
-                var avgReads = reader.IsDBNull(5) ? 0.0 : Convert.ToDouble(reader.GetValue(5));
-                var totalCpuUs = reader.IsDBNull(6) ? 0L : Convert.ToInt64(reader.GetValue(6));
-                var totalReads = reader.IsDBNull(7) ? 0L : Convert.ToInt64(reader.GetValue(7));
-                var totalSpills = reader.IsDBNull(8) ? 0L : Convert.ToInt64(reader.GetValue(8));
-                var maxDop = reader.IsDBNull(9) ? 0 : Convert.ToInt32(reader.GetValue(9));
-                var queryText = reader.IsDBNull(10) ? "" : reader.GetString(10);
-
-                // Skip low-impact queries — need meaningful per-execution cost
-                if (avgCpuMs < 10 && avgReads < 1000) continue;
-
-                facts.Add(new Fact
-                {
-                    Source = "bad_actor",
-                    Key = $"BAD_ACTOR_{queryHash}",
-                    Value = avgCpuMs, // Primary scoring dimension
-                    ServerId = context.ServerId,
-                    DatabaseName = dbName,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["execution_count"] = execCount,
-                        ["avg_cpu_ms"] = avgCpuMs,
-                        ["avg_elapsed_ms"] = avgElapsedMs,
-                        ["avg_reads"] = avgReads,
-                        ["total_cpu_us"] = totalCpuUs,
-                        ["total_reads"] = totalReads,
-                        ["total_spills"] = totalSpills,
-                        ["max_dop"] = maxDop
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectBadActorFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects key perfmon counters: Page Life Expectancy, Batch Requests/sec, compilations.
-    /// PLE is scored; others are throughput context for the AI.
-    /// </summary>
-    private async Task CollectPerfmonFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        counter_name,
-        cntr_value,
-        cntr_value_delta,
-        ROW_NUMBER() OVER (PARTITION BY counter_name ORDER BY collection_time DESC) AS rn
-    FROM collect.perfmon_stats
-    WHERE collection_time >= @startTime
-    AND   collection_time <= @endTime
-    AND   counter_name IN ('Page life expectancy', 'Batch Requests/sec', 'SQL Compilations/sec', 'SQL Re-Compilations/sec')
-)
-SELECT counter_name, cntr_value, cntr_value_delta
-FROM latest WHERE rn = 1";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            while (await reader.ReadAsync())
-            {
-                var counterName = reader.GetString(0);
-                var cntrValue = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-                var deltaValue = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-
-                var (factKey, source) = counterName switch
-                {
-                    "Page life expectancy" => ("PERFMON_PLE", "perfmon"),
-                    "Batch Requests/sec" => ("PERFMON_BATCH_REQ_SEC", "perfmon"),
-                    "SQL Compilations/sec" => ("PERFMON_COMPILATIONS_SEC", "perfmon"),
-                    "SQL Re-Compilations/sec" => ("PERFMON_RECOMPILATIONS_SEC", "perfmon"),
-                    _ => (null, null)
-                };
-
-                if (factKey == null) continue;
-
-                // For PLE, use the absolute value. For rate counters, use delta.
-                var value = counterName == "Page life expectancy" ? (double)cntrValue : (double)deltaValue;
-
-                facts.Add(new Fact
-                {
-                    Source = source!,
-                    Key = factKey,
-                    Value = value,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["cntr_value"] = cntrValue,
-                        ["delta_cntr_value"] = deltaValue
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectPerfmonFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects top memory clerks by size. Context for understanding where memory is allocated.
-    /// Dashboard stores pages_kb — convert to MB for consistency with Lite facts.
-    /// </summary>
-    private async Task CollectMemoryClerkFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        clerk_type,
-        SUM(pages_kb) / 1024.0 AS memory_mb,
-        ROW_NUMBER() OVER (PARTITION BY clerk_type ORDER BY collection_time DESC) AS rn,
-        collection_time
-    FROM collect.memory_clerks_stats
-    WHERE collection_time <= @endTime
-    GROUP BY clerk_type, collection_time
-)
-SELECT TOP 10 clerk_type, memory_mb
-FROM latest WHERE rn = 1 AND memory_mb > 0
-ORDER BY memory_mb DESC";
-
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            var metadata = new Dictionary<string, double>();
-            var totalMb = 0.0;
-            var clerkCount = 0;
-
-            while (await reader.ReadAsync())
-            {
-                var clerkType = reader.GetString(0);
-                var memoryMb = Convert.ToDouble(reader.GetValue(1));
-                metadata[clerkType] = memoryMb;
-                totalMb += memoryMb;
-                clerkCount++;
-            }
-
-            if (clerkCount == 0) return;
-
-            metadata["total_top_clerks_mb"] = totalMb;
-            metadata["clerk_count"] = clerkCount;
-
-            facts.Add(new Fact
-            {
-                Source = "memory",
-                Key = "MEMORY_CLERKS",
-                Value = totalMb,
-                ServerId = context.ServerId,
-                Metadata = metadata
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectMemoryClerkFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects database configuration facts: RCSI status, auto_shrink, auto_close,
-    /// recovery model. Aggregates counts across databases.
-    /// Dashboard stores config as individual setting rows in config.database_configuration_history.
-    /// We pivot from the per-setting rows into aggregated counts.
-    /// </summary>
-    private async Task CollectDatabaseConfigFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        database_name,
-        setting_name,
-        setting_value,
-        ROW_NUMBER() OVER (PARTITION BY database_name, setting_name ORDER BY collection_time DESC) AS rn
-    FROM config.database_configuration_history
-    WHERE setting_type = 'DATABASE_PROPERTY' /* collector writes 'DATABASE_PROPERTY' (install/39); 'database_option' matched 0 rows → DB_CONFIG fact never fired */
-    AND   database_name NOT IN ('master', 'msdb', 'model', 'tempdb')
-),
-pivoted AS (
-    SELECT
-        database_name,
-        MAX(CASE WHEN setting_name = 'recovery_model_desc' THEN CAST(setting_value AS NVARCHAR(128)) END) AS recovery_model,
-        MAX(CASE WHEN setting_name = 'is_auto_shrink_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_auto_shrink_on,
-        MAX(CASE WHEN setting_name = 'is_auto_close_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_auto_close_on,
-        MAX(CASE WHEN setting_name = 'is_read_committed_snapshot_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_read_committed_snapshot_on,
-        MAX(CASE WHEN setting_name = 'is_auto_create_stats_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_auto_create_stats_on,
-        MAX(CASE WHEN setting_name = 'is_auto_update_stats_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_auto_update_stats_on,
-        MAX(CASE WHEN setting_name = 'page_verify_option_desc' THEN CAST(setting_value AS NVARCHAR(128)) END) AS page_verify_option,
-        MAX(CASE WHEN setting_name = 'is_query_store_on' THEN CAST(setting_value AS NVARCHAR(10)) END) AS is_query_store_on
-    FROM latest
-    WHERE rn = 1
-    GROUP BY database_name
-)
-SELECT
-    COUNT(*) AS database_count,
-    COUNT(CASE WHEN is_auto_shrink_on = '1' OR is_auto_shrink_on = 'True' THEN 1 END) AS auto_shrink_count,
-    COUNT(CASE WHEN is_auto_close_on = '1' OR is_auto_close_on = 'True' THEN 1 END) AS auto_close_count,
-    COUNT(CASE WHEN is_read_committed_snapshot_on = '0' OR is_read_committed_snapshot_on = 'False' THEN 1 END) AS rcsi_off_count,
-    COUNT(CASE WHEN is_auto_create_stats_on = '0' OR is_auto_create_stats_on = 'False' THEN 1 END) AS auto_create_stats_off_count,
-    COUNT(CASE WHEN is_auto_update_stats_on = '0' OR is_auto_update_stats_on = 'False' THEN 1 END) AS auto_update_stats_off_count,
-    COUNT(CASE WHEN page_verify_option IS NOT NULL AND page_verify_option != 'CHECKSUM' THEN 1 END) AS page_verify_not_checksum_count,
-    COUNT(CASE WHEN recovery_model = 'FULL' THEN 1 END) AS full_recovery_count,
-    COUNT(CASE WHEN recovery_model = 'SIMPLE' THEN 1 END) AS simple_recovery_count,
-    COUNT(CASE WHEN is_query_store_on = '1' OR is_query_store_on = 'True' THEN 1 END) AS query_store_on_count
-FROM pivoted";
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var dbCount = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (dbCount == 0) return;
-
-            var autoShrink = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var autoClose = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var rcsiOff = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var autoCreateOff = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-            var autoUpdateOff = reader.IsDBNull(5) ? 0L : Convert.ToInt64(reader.GetValue(5));
-            var pageVerifyBad = reader.IsDBNull(6) ? 0L : Convert.ToInt64(reader.GetValue(6));
-            var fullRecovery = reader.IsDBNull(7) ? 0L : Convert.ToInt64(reader.GetValue(7));
-            var simpleRecovery = reader.IsDBNull(8) ? 0L : Convert.ToInt64(reader.GetValue(8));
-            var queryStoreOn = reader.IsDBNull(9) ? 0L : Convert.ToInt64(reader.GetValue(9));
-
-            facts.Add(new Fact
-            {
-                Source = "database_config",
-                Key = "DB_CONFIG",
-                Value = dbCount,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["database_count"] = dbCount,
-                    ["auto_shrink_on_count"] = autoShrink,
-                    ["auto_close_on_count"] = autoClose,
-                    ["rcsi_off_count"] = rcsiOff,
-                    ["auto_create_stats_off_count"] = autoCreateOff,
-                    ["auto_update_stats_off_count"] = autoUpdateOff,
-                    ["page_verify_not_checksum_count"] = pageVerifyBad,
-                    ["full_recovery_count"] = fullRecovery,
-                    ["simple_recovery_count"] = simpleRecovery,
-                    ["query_store_on_count"] = queryStoreOn
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectDatabaseConfigFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects the percent-autogrowth-on-large-files config fact (WS3): data/log files set
-    /// to grow in PERCENTAGE steps that are also large (>= 10 GB), where a single growth is a
-    /// huge, stalling allocation. Reads the latest snapshot per file from
-    /// collect.database_size_stats, excludes system databases, and emits ONE aggregate
-    /// FILE_AUTOGROWTH_PERCENT fact carrying the offending-file/database counts (the per-file
-    /// detail + copy-paste fix is attached later by the drill-down collector).
-    /// </summary>
-    private async Task CollectFileAutogrowthFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        database_name,
-        file_id,
-        total_size_mb,
-        is_percent_growth,
-        ROW_NUMBER() OVER (PARTITION BY database_name, file_id ORDER BY collection_time DESC) AS rn
-    FROM collect.database_size_stats
-    WHERE database_name NOT IN ('master', 'msdb', 'model', 'tempdb')
-)
-SELECT
-    file_count = COUNT(*),
-    database_count = COUNT(DISTINCT database_name)
-FROM latest
-WHERE rn = 1
-AND   is_percent_growth = 1
-AND   total_size_mb >= @minSizeMb;";
-
-            cmd.Parameters.Add(new SqlParameter("@minSizeMb", 10240.0)); /* 10 GB */
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var fileCount = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (fileCount == 0) return;
-
-            var databaseCount = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "FILE_AUTOGROWTH_PERCENT",
-                Value = fileCount,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["file_count"] = fileCount,
-                    ["database_count"] = databaseCount
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectFileAutogrowthFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects procedure stats: top procedure by delta CPU time in the period.
-    /// </summary>
-    private async Task CollectProcedureStatsFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    COUNT(DISTINCT object_name) AS distinct_procs,
-    SUM(execution_count_delta) AS total_executions,
-    SUM(total_worker_time_delta) AS total_cpu_time_us,
-    SUM(total_elapsed_time_delta) AS total_elapsed_time_us,
-    SUM(total_logical_reads_delta) AS total_logical_reads
-FROM collect.procedure_stats
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime
-AND   execution_count_delta > 0";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var distinctProcs = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            var totalExecs = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var totalCpuUs = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalElapsedUs = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var totalReads = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            if (totalExecs == 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "queries",
-                Key = "PROCEDURE_STATS",
-                Value = totalCpuUs,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["distinct_procedures"] = distinctProcs,
-                    ["total_executions"] = totalExecs,
-                    ["total_cpu_time_us"] = totalCpuUs,
-                    ["total_elapsed_time_us"] = totalElapsedUs,
-                    ["total_logical_reads"] = totalReads,
-                    ["avg_cpu_per_exec_us"] = totalExecs > 0 ? (double)totalCpuUs / totalExecs : 0
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectProcedureStatsFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects active query snapshot facts: long-running queries, blocked sessions, high DOP.
-    /// Dashboard query_snapshots table is created by sp_WhoIsActive dynamically.
-    /// We query it if it exists.
-    /// </summary>
-    private async Task CollectActiveQueryFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            // Check if the table exists first (created dynamically by sp_WhoIsActive)
-            using var checkCmd = connection.CreateCommand();
-            checkCmd.CommandText = "SELECT OBJECT_ID(N'collect.query_snapshots', N'U')";
-            var tableExists = await checkCmd.ExecuteScalarAsync();
-            if (tableExists == null || tableExists == DBNull.Value) return;
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    COUNT(*) AS total_snapshots,
-    COUNT(CASE WHEN DATEDIFF(MILLISECOND, 0, [elapsed_time]) > 30000 THEN 1 END) AS long_running_count,
-    COUNT(CASE WHEN [blocking_session_id] IS NOT NULL AND [blocking_session_id] != '' THEN 1 END) AS blocked_count,
-    MAX(DATEDIFF(MILLISECOND, 0, [elapsed_time])) AS max_elapsed_ms,
-    COUNT(DISTINCT [session_id]) AS distinct_sessions
-FROM collect.query_snapshots
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalSnapshots = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (totalSnapshots == 0) return;
-
-            var longRunning = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var blocked = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var maxElapsed = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var distinctSessions = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-
-            facts.Add(new Fact
-            {
-                Source = "queries",
-                Key = "ACTIVE_QUERIES",
-                Value = longRunning,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["total_snapshots"] = totalSnapshots,
-                    ["long_running_count"] = longRunning,
-                    ["blocked_count"] = blocked,
-                    ["max_elapsed_ms"] = maxElapsed,
-                    ["distinct_sessions"] = distinctSessions
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectActiveQueryFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects running job facts: jobs currently running long vs historical averages.
-    /// </summary>
-    private async Task CollectRunningJobFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT
-    COUNT(*) AS running_count,
-    COUNT(CASE WHEN is_running_long = 1 THEN 1 END) AS running_long_count,
-    MAX(percent_of_average) AS max_percent_of_avg,
-    MAX(current_duration_seconds) AS max_duration_seconds
-FROM collect.running_jobs
-WHERE collection_time >= @startTime
-AND   collection_time <= @endTime";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var runningCount = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (runningCount == 0) return;
-
-            var runningLong = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var maxPctAvg = reader.IsDBNull(2) ? 0.0 : Convert.ToDouble(reader.GetValue(2));
-            var maxDuration = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-
-            facts.Add(new Fact
-            {
-                Source = "jobs",
-                Key = "RUNNING_JOBS",
-                Value = runningLong,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["running_count"] = runningCount,
-                    ["running_long_count"] = runningLong,
-                    ["max_percent_of_average"] = maxPctAvg,
-                    ["max_duration_seconds"] = maxDuration
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectRunningJobFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects session stats: connection counts, total connections.
-    /// Dashboard session_stats is a flat table (not per-program_name), so we adapt.
-    /// </summary>
-    private async Task CollectSessionFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        total_sessions,
-        running_sessions,
-        sleeping_sessions,
-        dormant_sessions,
-        databases_with_connections,
-        top_application_connections,
-        ROW_NUMBER() OVER (ORDER BY collection_time DESC) AS rn
-    FROM collect.session_stats
-    WHERE collection_time >= @startTime
-    AND   collection_time <= @endTime
-)
-SELECT
-    total_sessions AS total_connections,
-    running_sessions AS total_running,
-    sleeping_sessions AS total_sleeping,
-    dormant_sessions AS total_dormant,
-    databases_with_connections AS distinct_apps,
-    top_application_connections AS max_app_connections
-FROM latest WHERE rn = 1";
-
-            cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var totalConns = reader.IsDBNull(0) ? 0L : Convert.ToInt64(reader.GetValue(0));
-            if (totalConns == 0) return;
-
-            var totalRunning = reader.IsDBNull(1) ? 0L : Convert.ToInt64(reader.GetValue(1));
-            var totalSleeping = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalDormant = reader.IsDBNull(3) ? 0L : Convert.ToInt64(reader.GetValue(3));
-            var distinctApps = reader.IsDBNull(4) ? 0L : Convert.ToInt64(reader.GetValue(4));
-            var maxAppConns = reader.IsDBNull(5) ? 0L : Convert.ToInt64(reader.GetValue(5));
-
-            facts.Add(new Fact
-            {
-                Source = "sessions",
-                Key = "SESSION_STATS",
-                Value = totalConns,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["total_connections"] = totalConns,
-                    ["total_running"] = totalRunning,
-                    ["total_sleeping"] = totalSleeping,
-                    ["total_dormant"] = totalDormant,
-                    ["distinct_applications"] = distinctApps,
-                    ["max_app_connections"] = maxAppConns
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectSessionFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects active global trace flags. Context for the AI to factor into recommendations.
-    /// </summary>
-    private async Task CollectTraceFlagFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        trace_flag,
-        status,
-        ROW_NUMBER() OVER (PARTITION BY trace_flag ORDER BY collection_time DESC) AS rn
-    FROM config.trace_flags_history
-    WHERE is_global = 1
-)
-SELECT trace_flag
-FROM latest WHERE rn = 1 AND status = 1
-ORDER BY trace_flag";
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            var metadata = new Dictionary<string, double>();
-            var flagCount = 0;
-
-            while (await reader.ReadAsync())
-            {
-                var flag = Convert.ToInt32(reader.GetValue(0));
-                metadata[$"TF_{flag}"] = 1;
-                flagCount++;
-            }
-
-            if (flagCount == 0) return;
-
-            metadata["flag_count"] = flagCount;
-
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "TRACE_FLAGS",
-                Value = flagCount,
-                ServerId = context.ServerId,
-                Metadata = metadata
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectTraceFlagFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
     /// Collects server hardware properties: CPU count, cores, sockets, memory.
     /// Critical context for MAXDOP and memory recommendations.
     /// </summary>
@@ -2007,249 +104,6 @@ SELECT TOP 1
     product_version{health}
 FROM collect.server_properties
 ORDER BY collection_time DESC";
-    }
-
-    private async Task CollectServerPropertiesFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            // Version-skew resilience: a server whose PerformanceMonitor DB has not yet had the WS5
-            // upgrade lacks some or all of the three server-health columns. Probe EACH independently
-            // (COL_LENGTH returns NULL for an absent column) and reference only the present ones, so
-            // the core SERVER_HARDWARE read never fails — and keeps flowing — regardless of which
-            // columns a partially-upgraded or out-of-order schema happens to have.
-            bool hasLpim, hasIfi, hasDumps;
-            using (var probe = connection.CreateCommand())
-            {
-                probe.CommandText = $@"
-SELECT
-    COL_LENGTH('collect.server_properties', '{LpimColumn}'),
-    COL_LENGTH('collect.server_properties', '{IfiColumn}'),
-    COL_LENGTH('collect.server_properties', '{DumpsColumn}');";
-                using var probeReader = await probe.ExecuteReaderAsync();
-                await probeReader.ReadAsync();
-                hasLpim = !probeReader.IsDBNull(0);
-                hasIfi = !probeReader.IsDBNull(1);
-                hasDumps = !probeReader.IsDBNull(2);
-            }
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = BuildServerPropertiesQuery(hasLpim, hasIfi, hasDumps);
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var cpuCount = reader.IsDBNull(0) ? 0 : Convert.ToInt32(reader.GetValue(0));
-            var htRatio = reader.IsDBNull(1) ? 0 : Convert.ToInt32(reader.GetValue(1));
-            var physicalMemMb = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var socketCount = reader.IsDBNull(3) ? 0 : Convert.ToInt32(reader.GetValue(3));
-            var coresPerSocket = reader.IsDBNull(4) ? 0 : Convert.ToInt32(reader.GetValue(4));
-            var hadrEnabled = !reader.IsDBNull(5) && Convert.ToBoolean(reader.GetValue(5));
-            var edition = reader.IsDBNull(6) ? string.Empty : reader.GetString(6);
-
-            // Read each PRESENT health column by name — the SELECT includes only the columns that
-            // exist, so their ordinals shift with the subset; GetOrdinal resolves each regardless.
-            // An absent column stays null, so EmitServerHealthFacts emits nothing for it.
-            bool? lpim = null;
-            bool? ifi = null;
-            int? dumpCount = null;
-            if (hasLpim)
-            {
-                var ord = reader.GetOrdinal(LpimColumn);
-                lpim = reader.IsDBNull(ord) ? (bool?)null : Convert.ToBoolean(reader.GetValue(ord));
-            }
-            if (hasIfi)
-            {
-                var ord = reader.GetOrdinal(IfiColumn);
-                ifi = reader.IsDBNull(ord) ? (bool?)null : Convert.ToBoolean(reader.GetValue(ord));
-            }
-            if (hasDumps)
-            {
-                var ord = reader.GetOrdinal(DumpsColumn);
-                dumpCount = reader.IsDBNull(ord) ? (int?)null : Convert.ToInt32(reader.GetValue(ord));
-            }
-
-            if (cpuCount == 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "config",
-                Key = "SERVER_HARDWARE",
-                Value = cpuCount,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["cpu_count"] = cpuCount,
-                    ["hyperthread_ratio"] = htRatio,
-                    ["physical_memory_mb"] = physicalMemMb,
-                    ["socket_count"] = socketCount,
-                    ["cores_per_socket"] = coresPerSocket,
-                    ["hadr_enabled"] = hadrEnabled ? 1 : 0
-                }
-            });
-
-            // WS5 server-health advisories (advise-only). Gating lives here so a fact that would
-            // score 0 is simply never emitted (noise control); the scorer then scores the emitted
-            // fact's Value. Shared with Lite — keep the rules identical (see DuckDbFactCollector).
-            FactCollectorHelpers.EmitServerHealthFacts(context, facts, edition, physicalMemMb, lpim, ifi, dumpCount);
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectServerPropertiesFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// WS4: plan-XML advisories. Parses the already-collected query plans of the top queries by
-    /// cost (no live fetch, no DMV) with the shared ShowPlanParser/PlanAnalyzer and emits two
-    /// advise-only facts — MISSING_INDEX (Value = distinct suggested indexes) and PLAN_WARNING
-    /// (Value = actionable warnings). The specific indexes/warnings ride in the finding drill-down
-    /// (SqlServerDrillDownCollector); Fact.Metadata is numeric only.
-    /// </summary>
-    private async Task CollectPlanAdvisoryFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            var planXmls = new List<string>();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync();
-
-                using var cmd = connection.CreateCommand();
-                cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-SELECT TOP (10)
-    plan_xml = CAST(DECOMPRESS(qs.query_plan_text) AS nvarchar(max))
-FROM collect.query_stats AS qs
-WHERE qs.collection_time >= @startTime
-AND   qs.collection_time <= @endTime
-AND   qs.query_plan_text IS NOT NULL
-ORDER BY
-    qs.total_worker_time DESC;";
-                cmd.Parameters.Add(new SqlParameter("@startTime", context.TimeRangeStart));
-                cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-                using var reader = await cmd.ExecuteReaderAsync();
-                while (await reader.ReadAsync())
-                {
-                    if (!reader.IsDBNull(0))
-                        planXmls.Add(reader.GetString(0));
-                }
-            }
-
-            if (planXmls.Count == 0)
-                return;
-
-            var summary = PlanAdvisoryAggregator.Summarize(planXmls);
-
-            if (summary.MissingIndexCount > 0)
-            {
-                facts.Add(new Fact
-                {
-                    Source = "queries",
-                    Key = "MISSING_INDEX",
-                    Value = summary.MissingIndexCount,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["index_count"] = summary.MissingIndexCount,
-                        ["max_impact"] = summary.MaxImpact
-                    }
-                });
-            }
-
-            if (summary.WarningCount > 0)
-            {
-                facts.Add(new Fact
-                {
-                    Source = "queries",
-                    Key = "PLAN_WARNING",
-                    Value = summary.WarningCount,
-                    ServerId = context.ServerId,
-                    Metadata = new Dictionary<string, double>
-                    {
-                        ["warning_count"] = summary.WarningCount,
-                        ["critical_count"] = summary.CriticalCount
-                    }
-                });
-            }
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectPlanAdvisoryFactsAsync failed", ex);
-        }
-    }
-
-    /// <summary>
-    /// Collects disk space facts from database_size_stats: volume free space, file sizes.
-    /// </summary>
-    private async Task CollectDiskSpaceFactsAsync(AnalysisContext context, List<Fact> facts)
-    {
-        try
-        {
-            using var connection = new SqlConnection(_connectionString);
-            await connection.OpenAsync();
-
-            using var cmd = connection.CreateCommand();
-            cmd.CommandText = @"
-SET TRANSACTION ISOLATION LEVEL READ UNCOMMITTED;
-
-;WITH latest AS (
-    SELECT
-        volume_mount_point,
-        volume_total_mb,
-        volume_free_mb,
-        ROW_NUMBER() OVER (PARTITION BY volume_mount_point ORDER BY collection_time DESC) AS rn
-    FROM collect.database_size_stats
-    WHERE collection_time <= @endTime
-    AND   volume_total_mb > 0
-)
-SELECT
-    MIN(volume_free_mb * 1.0 / volume_total_mb) AS min_free_pct,
-    MIN(volume_free_mb) AS min_free_mb,
-    COUNT(DISTINCT volume_mount_point) AS volume_count,
-    SUM(volume_total_mb) AS total_volume_mb,
-    SUM(volume_free_mb) AS total_free_mb
-FROM latest WHERE rn = 1";
-
-            cmd.Parameters.Add(new SqlParameter("@endTime", context.TimeRangeEnd));
-
-            using var reader = await cmd.ExecuteReaderAsync();
-            if (!await reader.ReadAsync()) return;
-
-            var minFreePct = reader.IsDBNull(0) ? 1.0 : Convert.ToDouble(reader.GetValue(0));
-            var minFreeMb = reader.IsDBNull(1) ? 0.0 : Convert.ToDouble(reader.GetValue(1));
-            var volumeCount = reader.IsDBNull(2) ? 0L : Convert.ToInt64(reader.GetValue(2));
-            var totalVolumeMb = reader.IsDBNull(3) ? 0.0 : Convert.ToDouble(reader.GetValue(3));
-            var totalFreeMb = reader.IsDBNull(4) ? 0.0 : Convert.ToDouble(reader.GetValue(4));
-
-            if (volumeCount == 0) return;
-
-            facts.Add(new Fact
-            {
-                Source = "disk",
-                Key = "DISK_SPACE",
-                Value = minFreePct,
-                ServerId = context.ServerId,
-                Metadata = new Dictionary<string, double>
-                {
-                    ["min_free_pct"] = minFreePct,
-                    ["min_free_mb"] = minFreeMb,
-                    ["volume_count"] = volumeCount,
-                    ["total_volume_mb"] = totalVolumeMb,
-                    ["total_free_mb"] = totalFreeMb
-                }
-            });
-        }
-        catch (Exception ex)
-        {
-            Logger.Error("SqlServerFactCollector.CollectDiskSpaceFactsAsync failed", ex);
-        }
     }
 
 }
